fix(engine-v2): stage-1 fixes from adversarial review (7 unique defects)
- [major] spares feature double-subtracted the head stack whenever the head
  cell (own 24 = enemy frame 12) sat inside the strongest block — a class
  measured at ~50% of contact positions; now the exclusion is the union
  (head subtracted separately only when outside the block)
- block-derived features (front edge, trapped, depth) gated to runs of 2+;
  a lone occupied cell no longer reports trapped/depth with a hidden edge
- depth normalized by 23 (was /18, could reach 1.28)
- free_pips commits ONE checker per block cell (stacked spares on block
  points move freely — their pips are timing, per the documented semantics)
- net2 from_bytes: total-size check in u64 — on wasm32 a crafted 64×4096²
  header wrapped usize and passed validation
- train_v2.py: empty NNF2 file fails fast BEFORE training the other net;
  --val 0 now genuinely disables validation (max(1,...) used to hold out
  one row and key best-model selection to a single sample's CE)
- broken test fixture rebuilt (Black placement overwrote a prime cell making
  the trapped assertion vacuous) + 3 regression tests for the above

86 engine-core tests (3 new); trainer smoke: --val 0 and empty-race paths
verified.

+++ crates/engine-core/src/net2.rs
@@ -159,8 +159,11 @@ impl NetV2 {
         if dims.windows(2).any(|d| d[0].1 != d[1].0) {
             return None;
         }
-        let total: usize = dims.iter().map(|&(i, o)| i * o + o).sum();
-        if bytes.len() != off + total * 4 {
+        // u64 arithmetic: 64 layers × 4096² floats × 4 bytes overflows usize
+        // on wasm32, where this loader also runs — a crafted header must be
+        // rejected, not wrap into a "valid" length and trap later.
+        let total: u64 = dims.iter().map(|&(i, o)| (i as u64) * (o as u64) + o as u64).sum();
+        if bytes.len() as u64 != off as u64 + total * 4 {
             return None;
         }
         let mut layers = Vec::with_capacity(n);
